% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{load_from_file}
\alias{load_from_file}
\title{Loads a binary protobuf that stores onnx model}
\usage{
load_from_file(obj)
}
\arguments{
\item{obj}{a file-like object (has "read" function)
or a string containing a file name}
}
\value{
ONNX ModelProto object
}
\description{
Loads a binary protobuf that stores onnx model
}
