% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{load_from_string}
\alias{load_from_string}
\title{Loads a binary string that stores onnx model}
\usage{
load_from_string(s)
}
\arguments{
\item{a}{string object containing protobuf}
}
\value{
ONNX ModelProto object
}
\description{
Loads a binary string that stores onnx model
}
