% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maker.R
\name{make_attribute}
\alias{make_attribute}
\alias{make_attribute}
\title{Makes an AttributeProto based on the value type.}
\usage{
make_attribute(key, value, doc_string = NULL)

make_attribute(key, value, doc_string = NULL)
}
\arguments{
\item{key}{key}

\item{value}{value}

\item{doc_string}{doc_string}

\item{key}{key}

\item{value}{value}

\item{doc_string}{doc_string}
}
\description{
Makes an AttributeProto based on the value type.

Makes an AttributeProto based on the value type.
}
