% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maker.R
\name{make_graph}
\alias{make_graph}
\title{Makes an GraphProto.}
\usage{
make_graph(nodes, name, inputs, outputs, initializer = NULL,
  doc_string = NULL)
}
\arguments{
\item{nodes}{nodes}

\item{name}{name}

\item{inputs}{inputs}

\item{outputs}{outputs}

\item{initializer}{initializer}

\item{doc_string}{doc_string}
}
\description{
Makes an GraphProto.
}
