% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maker.R
\name{make_node}
\alias{make_node}
\title{Makes an NodeProto.}
\usage{
make_node(op_type, inputs, outputs, name = NULL, doc_string = NULL)
}
\arguments{
\item{op_type}{op_type}

\item{inputs}{inputs}

\item{outputs}{outputs}

\item{name}{name}

\item{doc_string}{doc_string}
}
\description{
Makes an NodeProto.
}
