% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maker.R
\name{make_tensor}
\alias{make_tensor}
\title{Make a TensorProto with specified arguments.}
\usage{
make_tensor(name, data_type, dims, vals, raw = FALSE)
}
\arguments{
\item{name}{name}

\item{data_type}{data_type}

\item{dims}{dims}

\item{vals}{vals}

\item{raw}{raw}
}
\description{
If raw is FALSE, this function will choose the corresponding proto field to
store the values based on data_type. If raw is TRUE, use "raw_data" proto
field to store the values, and values should be of type bytes in this case.
}
