% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maker.R
\name{make_tensor_value_info}
\alias{make_tensor_value_info}
\title{Makes a TypeProto based on the data type and shape.}
\usage{
make_tensor_value_info(name, elem_type, shape, doc_string = "")
}
\arguments{
\item{name}{name}

\item{elem_type}{elem_type}

\item{shape}{shape}

\item{doc_string}{doc_string}
}
\description{
Makes a TypeProto based on the data type and shape.
}
