% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{onnx-package}
\alias{onnx}
\alias{onnx-package}
\title{onnx: R Interface to 'ONNX'}
\description{
R Interface to 'ONNX' - Open Neural Network Exchange <https://onnx.ai/>. 
             'ONNX' provides an open source format for machine learning models. 
             It defines an extensible computation graph model, as well as definitions
             of built-in operators and standard data types.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/onnx/onnx-r}
  \item Report bugs at \url{https://github.com/onnx/onnx-r/issues}
}

}
\author{
\strong{Maintainer}: Yuan Tang \email{terrytangyuan@gmail.com} (\href{https://orcid.org/0000-0001-5243-233X}{ORCID})

Authors:
\itemize{
  \item ONNX Authors [copyright holder]
}

Other contributors:
\itemize{
  \item Facebook, Inc. [copyright holder]
  \item Microsoft Corporation [copyright holder]
}

}
\keyword{internal}
