% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print_readable}
\alias{print_readable}
\title{Print the Human-readable Representation of the Proto Object}
\usage{
print_readable(x)
}
\arguments{
\item{x}{The proto object}
}
\description{
Print the Human-readable Representation of the Proto Object
}
